Add hot-reload for server side
