+++ api/data/.env.example
@@ -1,2 +1,3 @@
 NODE_ENV=development
 DB_DEBUG=false
+PORT=8081
