+++ api/data/package.json
@@ -9,7 +9,7 @@
   },
   "scripts": {
     "ci": "yarn lint && yarn test",
-    "start": "cross-env NODE_ENV=development babel-node src | bunyan -o short --color",
+    "start": "webpack | bunyan -o short --color",
     "test": "jest --config jest.config.js --coverage",
     "test:watch": "jest --config jest.config.js --watch",
     "lint": "eslint . --ignore-path ../../.gitignore",
@@ -31,13 +31,10 @@
     "axios": "^0.18.0",
     "bluebird": "^3.5.1",
     "bunyan": "^1.8.12",
-    "cross-env": "^5.1.3",
     "dotenv": "^5.0.1",
     "fs-extra": "^5.0.0",
     "graphql": "^0.13.2",
-    "graphql-tools": "^2.21.0",
     "knex": "^0.14.4",
-    "koa-bunyan-logger": "^2.0.0",
     "lodash": "^4.17.5",
     "sanitize-filename": "^1.6.1",
     "sqlite3": "^3.1.13"
@@ -47,9 +44,10 @@
     "babel-core": "^6.26.0",
     "babel-eslint": "^8.2.2",
     "babel-jest": "^22.4.1",
+    "babel-loader": "^7.1.5",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-bluebird": "^1.0.1",
-    "babel-preset-env": "^1.6.1",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-flow": "^6.23.0",
     "eslint": "^4.18.2",
     "eslint-config-airbnb-base": "^12.1.0",
@@ -57,7 +55,11 @@
     "eslint-import-resolver-node": "^0.3.2",
     "eslint-plugin-import": "^2.9.0",
     "eslint-plugin-prettier": "^2.6.0",
-    "jest": "^22.4.2",
-    "prettier": "^1.11.1"
+    "jest": "^23.5.0",
+    "prettier": "^1.14.2",
+    "start-server-webpack-plugin": "^2.2.5",
+    "webpack": "^4.17.1",
+    "webpack-cli": "^3.1.0",
+    "webpack-node-externals": "^1.7.2"
   }
 }
